test(tree): fix test unit about drag drop #INFR-11456
